[master] updated scripts

+++ bin/buildall.sh
@@ -81,7 +81,7 @@ coverage=( "*.gcda" "*.gcno" "*.s" "*.i" "*.o" "*.info" )
 directories=( "build-v1" "build-v2" "build-v3" "build-v4" "build-v5" "build-v6" )
 for directory in ${directories[@]}; do
     for f in ${coverage}; do
-        find "${directory}" -type f -name "${f}" -delete
+        [[ -d "${directory}" ]] && find "${directory}" -type f -name "${f}" -delete
     done
 done
